requirements.txt
together>=1.3.0
jinja2>=3.1.0
pyyaml>=6.0
requests>=2.31.0
edge-tts>=7.2.7
playwright>=1.43.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.0
sse-starlette>=2.1.3
Pillow>=10.0.0
